services/ml-service/requirements.txt
# ============= CORE DEPENDENCIES =============
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ============= AWS SDK =============
boto3==1.34.0
botocore==1.34.0

# ============= TESTING =============
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
moto[all]==4.2.9

# ============= ML DEPENDENCIES =============
# For stub/local inference
numpy==1.26.2
pillow==10.1.0

# Uncomment for real TFLite inference:
# tensorflow==2.15.0  # CPU version
# tensorflow[and-cuda]==2.15.0  # GPU version
# opencv-python-headless==4.8.1.78

# ============= UTILITIES =============
python-dotenv==1.0.0
